backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.23.1
youtube-transcript-api==0.6.0
httpx==0.25.0
python-dotenv==1.0.0
pydantic==2.8.0
google-genai
